+++ test/tx-data/sell-all-swap-pool.test.js
@@ -21,4 +21,16 @@ describe('TxDataSellAllSwapPool', () => {
     test('TxData', () => {
         expect(TxData(params, TX_TYPE.SELL_ALL_SWAP_POOL).serialize()).toEqual(serializedTxData);
     });
+
+    test('empty data', () => {
+        const txData = new TxDataSellAllSwapPool();
+        const emptySerializedTxData = (txData).serialize();
+
+        expect(emptySerializedTxData.toString('hex')).toEqual('c1c0');
+
+        expect(decodeToArray(emptySerializedTxData))
+            .toEqual([
+                [192],
+            ]);
+    });
 });